IntegraHub/services/demo_portal/requirements.txt
fastapi
uvicorn
jinja2
httpx
python-multipart
# Dependencias compartidas
pika
pybreaker
tenacity
PyJWT
sqlalchemy
psycopg2-binary
